backend/worker/video_generation/requirements.txt
psycopg2-binary
numpy
imageio
imageio-ffmpeg
google-cloud-storage
google-genai
argparse